+++ test/managedTests.hs
@@ -72,11 +72,11 @@ makeWithWrapperWithMessage ref inMsg outMsg v handler = do
 
 managedCauldron :: IORef [Text] -> Cauldron Managed
 managedCauldron ref =
-  fromRecipeList
+  mconcat
     [ recipe @(Logger IO) $ eff $ wire $ managed (makeLogger ref),
       recipe @(Weird IO)
         Recipe
-          { bean = eff do
+          { bare = eff do
               wire \logger self -> managed (makeSelfInvokingWeird ref logger self),
             decos =
               fromDecoList